Changed delete step behavior to allow for non-existent paths
Before this would cause an assert, stacktrace and abort of the program.

Now it causes a warning:

WARNING: delete: path 'out/steps/001_masterfiles_793214c10cd0d1f31e8ca44ab05790554776772a/templates/README.md' does not exist, skipping

Ticket: CFE-4257
Changelog: title

+++ cfbs/utils.py
@@ -5,6 +5,7 @@ import json
 import copy
 import subprocess
 import hashlib
+import logging as log
 import urllib
 import urllib.request  # needed on some platforms
 from collections import OrderedDict
@@ -56,10 +57,13 @@ def touch(path: str):
 def rm(path: str, missing_ok=False):
     if not missing_ok:
         assert os.path.exists(path)
+    if missing_ok and not os.path.exists(path):
+        return False
     if os.path.isdir(path):
         rmtree(path)
     if os.path.isfile(path):
         os.remove(path)
+    return True
 
 
 def cp(src, dst):
